requirements.txt
pillow==11.0.0
numpy==2.2.0
matplotlib==3.10.0